+++ Userland/Libraries/LibJS/Heap/Cell.h
@@ -14,8 +14,6 @@
 #include <LibJS/Forward.h>
 #include <LibJS/Heap/GCPtr.h>
 #include <LibJS/Heap/Internals.h>
-#include <LibJS/Runtime/Completion.h>
-#include <LibJS/Runtime/Value.h>
 
 namespace JS {
 
@@ -33,7 +31,7 @@ class Cell {
     AK_MAKE_NONMOVABLE(Cell);
 
 public:
-    virtual ThrowCompletionOr<void> initialize(Realm&) { return {}; }
+    virtual ThrowCompletionOr<void> initialize(Realm&);
     virtual ~Cell() = default;
 
     bool is_marked() const { return m_mark; }
@@ -75,11 +73,7 @@ public:
             visit_impl(const_cast<RemoveConst<T>&>(*cell.ptr()));
         }
 
-        void visit(Value value)
-        {
-            if (value.is_cell())
-                visit_impl(value.as_cell());
-        }
+        void visit(Value value);
 
         // Allow explicitly ignoring a GC-allocated member in a visit_edges implementation instead
         // of just not using it.
