+++ Userland/Libraries/LibJS/Heap/Heap.h
@@ -21,6 +21,7 @@
 #include <LibJS/Heap/Handle.h>
 #include <LibJS/Heap/Internals.h>
 #include <LibJS/Heap/MarkedVector.h>
+#include <LibJS/Runtime/Completion.h>
 #include <LibJS/Runtime/WeakContainer.h>
 
 namespace JS {
