+++ Userland/Libraries/LibWeb/CSS/CSSKeyframeRule.cpp
@@ -11,6 +11,11 @@
 
 namespace Web::CSS {
 
+WebIDL::ExceptionOr<JS::NonnullGCPtr<CSSKeyframeRule>> CSSKeyframeRule::create(JS::Realm& realm, CSS::Percentage key, Web::CSS::CSSStyleDeclaration& declarations)
+{
+    return MUST_OR_THROW_OOM(realm.heap().allocate<CSSKeyframeRule>(realm, realm, key, declarations));
+}
+
 void CSSKeyframeRule::visit_edges(Visitor& visitor)
 {
     Base::visit_edges(visitor);
