+++ Userland/Libraries/LibWeb/DOM/IDLEventListener.cpp
@@ -4,6 +4,8 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <LibJS/Heap/Heap.h>
+#include <LibJS/Runtime/Error.h>
 #include <LibWeb/DOM/IDLEventListener.h>
 
 namespace Web::DOM {
