+++ Userland/Libraries/LibWeb/HTML/StructuredSerialize.cpp
@@ -8,6 +8,7 @@
 #include <AK/HashTable.h>
 #include <AK/Vector.h>
 #include <LibJS/Forward.h>
+#include <LibJS/Runtime/VM.h>
 #include <LibWeb/HTML/StructuredSerialize.h>
 #include <LibWeb/WebIDL/ExceptionOr.h>
