+++ Userland/Libraries/LibWeb/Painting/GradientPainting.cpp
@@ -6,6 +6,7 @@
 
 #include <AK/Math.h>
 #include <LibGfx/Gradients.h>
+#include <LibGfx/Painter.h>
 #include <LibWeb/CSS/StyleValues/ConicGradientStyleValue.h>
 #include <LibWeb/CSS/StyleValues/LinearGradientStyleValue.h>
 #include <LibWeb/CSS/StyleValues/RadialGradientStyleValue.h>
